fix: Don't crash on relationship severance events
Previous code was missing the JSON name of the event, so it wasn't
present when it should have been, resulting in a crash.

Also, the Mastodon documentation is incorrect about the relationship
count, which is instead represented as two properties, one for followers
and one for following. So model that, and display them separately in
the UI.

Fixes #1086