feat: Show extra a11y actions for trending links and suggested accounts
Extend the "suggested accounts" accessibility actions to include any
mentions in the account's bio. Links, mentions, and hashtags are now
shown with a button to easily copy them.

Extend the "trending links" accessibility actions with a new "copy
link" action.

Consolidate common functionality in to the new
`PachliRecyclerviewAccessibilityDelegate` base class.