feat: Show link title when viewing posts about a trending link
Previous code inadvertently crashed when the user clicekd on a trending
link count to see statuses about the link.

Don't do that. Instead, show the statuses that mention the link, and
show the link's title in the actionbar to make it more explicit for the
user.

Special-case this timeline type in TimelineActivity so it can't be
added to a tab (it would be difficult to distinguish it amongst tabs
as they would have the same icon).