refactor: Replace the different network response types with ApiResult (#1261)
Previous code used five (!) different types for the network response.

Some used Retrofit's `Response`. This provides access to the headers.

Some used `NetworkResult`. This did not provide access to the headers,
but did provide some higher-order functions (e.g., `fold`) for operating
on the results.

One used a raw `Map`.

One used a raw `Call`.

The rest had been converted to `ApiResult`, a `Result<V, ApiError>`.
This provides the higher-order functions, provides the headers, and
is exception-free, so is the correct type to use.

This PR completes the work of cutting over to `ApiResult`. The return
values are changed and the calling code is adjusted to use the new
functions as appropriate.