feat: Always show the post beng replied to when replying (#1263)
Previous code showed the post being replied to, but in a tiny font and
hidden behind a disclosure triangle that was difficult to spot.

Fix that. Show the post, including the author and their avatar. The text
of the post is selectable for copy/paste in to the reply if necessary
(links and hashtags are deliberately not clickable so they can't be
clicked by accident).

To do this:

Update `ComposeOptions`. Instead of three different properties that
were set if the post was a reply, use a new `InReplyTo` sealed class
that covers two situations; either the caller has the full content of
the status being replied to, in which case it's included, or they only
have the ID of the status being replied to.

Update `ComposeViewModel` with a state flow of `inReplyTo` results to
represent either (a) Not a reply, (b) is a reply, data is loading, (c)
is a reply, data is loaded, or (d) is a reply, error occurred loading
data.

In `ComposeActivity` use this flow to drive updates to a new part of
the UI showing the status being replied to (or hiding that part of the
UI if this is not a reply).