refactor: Simplify and centralise updates to StatusViewDataEntity (#1289)
Previous code considered all non-primary-key properties of
`StatusViewDataEntity` to be non-nullable.

This made setting the values more difficult than it needed to be, as a
complete `StatusViewDataEntity` needed to be created first.

Make these properties nullable, so `@Upsert` with partial data POJOs can
be used to update them.

Centralise the update code in `StatusRepository`. This also removes some
duplicate code.