feat: Implement content and account filters for conversations (#1293)
The user can now specify content and account filters for conversations.
Content filters apply to the most recent status in the conversation,
account filters apply to the account that initiated the thread. Once
the user replies they are ignored.

- Update the content filter settings to make it clear which filter
applies to Conversations.

- Add a new set of account preferences and UI allowing the user to
choose which account filters should apply to conversations, and store
these in `AccountEntity`. Account filters are only enabled if the
user's server supports the API call to fetch multiple statuses in a
single request.

- Store the conversation account filter preferences in the on-device
`PachliAccount`.

- Extend `ConversationAdapter` to show the conversation, a 
content-filtered version of the conversation, or an account-filtered
version.

As normal, when displaying content filtered with a warning the user has
the choice to show the content or edit the filters. Hidden content is
hidden with no disclosure.