feat: Add Atkinson Hyperlegible Next as a selectable font (#1311)
From https://www.brailleinstitute.org/freefont:

> Atkinson Hyperlegible Next is an enhanced font introduced in 2025.

According to
https://braileinstitute.app.box.com/s/rin3vzegmcy7sil28yfqslz2r5etv5nl
(the license download link) the license is the SIL Open Font License,
Version 1.1, and it appears the original Atkinson Hyperlegible is also
now available under the same license (in addition, the link to the
previous license is now a 404). So adjust the license info accordingly.