fix: Highlight modified hashtags when viewing status edits
Previous code used `unicodeWrap` to add the directional isolates, which
returns a `String`. This overwrites any existing spans in the text, and
in particular, the spans that marked insertions or deletions.

The effect was to remove the insert/delete markup for hashtags (and
probably mentions) when viewing edits to a status.

Fix this by avoiding `unicodeWrap` in this case and adding the isolates
to the spanned text.

While I'm here, cleanup up spoiler text wrangling in edits.