fix: Remove old account data before creating FK constraints
The 10 -> 11 database migration adds FK constraints with
`AccountEntity.id` as the parent.

Because of previous bugs there may be orphaned data in some tables that
references accounts that no longer exist in `AccountEntity`. This causes
a crash during migration.

This migration was previously automatic. Rewrite it to be manual, by
copying the code that was previously generated and inserting explcit
`DELETE` statements ahead of the migration to remove the orphaned rows.

Fixes #1320