fix: Invalidate NetworkTimelineRepository after clearing status filter (#1324)
When a status is partially filtered ("Hide with a warning"), its "Show
anyway" button would be no-op on some timelines, like "Trending posts"
or a user's profile.

It turns out the functionality was broken on
`NetworkTimelineViewModel/Repository`. The call to
`NetworkTimelineViewModel.clearWarning()` wasn't having an effect
because `repository.invalidate()` wasn't being called, although it was
present on all other methods that also made use of
`repository.updateActionableStatusById()` (`handleFavEvent()`,
`handleBookmarkEvent()` and `handlePinEvent()`).

Instead of just adding `repository.invalidate()` to `clearWarning()`, I
preferred to remove the call from the other methods and move it inside
`NetworkTimelineRepository.updateActionableStatusById()`. This should
prevent something similar from happening again, and is also how
`NetworkTimelineRepository.updateStatusById()` works.