fix: Ensure all languages are selectable via system settings (#1327)
`locales_config.xml` was out of sync with the available languages, which
meant that while some of them were selectable in the Pachli UI they
weren't selectable in the language system settings Android presents for
each app.

Fix this by updating `mklanguages` to also populate `locales_config.xml`
with the available languages.

This adds languages:

- Ελληνικά / Greek
- Suomi / Finnish
- Frysk / Western Frisian
- Indonesia / Indonesian
- Latviešu / Latvian
- മലയാളം / Malayalam
- සිංහල / Sinhala
- Slovenčina / Slovak

Fixes #1323