feat: Implement Markdown rendering, controlled by a lab preference (#1353)
Provide a new lab preference to control whether status content should be
pre-processed from HTML to Markdown, and then rendered as Markdown.

Hoist processing of status content out of the different viewholders into
implementations of a new `SetStatusContent` interface.

Provide two implementations of this interface; one processes the content
as Mastodon HTML, consistent with the previous behaviour.

The other implementation pre-processes the content to remove some HTML
block and other elements, then runs the content through the Markwon
Markdown processor.

The correct implementation has to be passed to the viewholders, so
modify the different fragments that show statuses to create the
appropriate implementation based on the user's preferences, and then
pass it through the adapter which will pass it to the viewholder.

Add this as a parameter to `StatusBaseViewHolder()`

Fixes #878