change: Render the privacy policy from Markdown at runtime
Previous code rendered the privacy policy to an HTML string resource at
build time using a custom plugin.

Now Pachli includes a Markdown parser it's possible to bundle the policy
as an asset and render it from Markdown at runtime instead.

This simplifies the build. It also makes the display of the privacy
policy consistent with the display of other Markdown content within
the app.

This is not dependent on the "Render Markdown" preference.

To make this work the privacy policy is copied to the assets directory
ahead of the `preBuild` step.