fix: Show join date in grammatically correct form (#1367)
Previous code used "MMMM" for the month, which is only appropriate if
the day of the month is also included (this is a "context sensitive"
month). While this looks correct in English it is incorrect in some
other languages.

Use "LLLL", the "standalone" form, which is correct in all languages
when the day of the month is not included.

Fixes #1358