fix: Support editing captions of uploaded images (#1382)
Previous code tried to use the media update API to update captions when
editing a status, but that only works if the status has not been posted
yet.

When editing a status the caption edits go through the status edit API,
so just record the new caption and exit early.

Fixes #1380