feat: Make new content filters default to active in all contexts
Content filters make sense if they're applied to all contexts.

Otherwise you can have a situation where e.g., a status that @-mentions
you is filtered from the Home timeline, but the notification about the
mention is not filtered from the Notifications timeline.

To make this easier for users change the default behaviour for the
filter contexts for new filters. Previously all contexts were disabled
and the user would choose which contexts to enable.

Now all contexts are enabled, and the user chooses which ones to
disable.

Fixes #1383