fix: Create filters with the correct content-type (#1390)
Content filters are created with a form-encoded POST request. The
Mastodon API defines these in a way that can't be replicated with
Retrofit's normal mechanisms for defining an API (including using the
`@FormUrlEncoded annotation`), so `NewContentFilterConverterFactory`
converts the `NewContentFilter` object to the body of the request.

However, for some servers (at least, GoToSocial) the content-type header
is required. That's normally provided by the `@FormUrlEncoded`
annotation, but that can't be used here.

So provide the header directly in the API definition.

Fixes #1375