refactor: Modularise the translation code
Abstract the translation service behind an interface to allow different
translation services in the future.

Provide core.model classes to represent a translation result, and mappers
to go to/from the network/entity classes and the model class.

Surface translation errors in ViewThreadViewModel so they are visible to
the user.