feat: Support translation throughout the app (#1414)
Previous code could translate statuses in the home timeline and threads,
but nowhere else.

Extend this support to almost all timelines, in particular:

- Notifications
- Local and federated timelines
- Trending
- Bookmarks
- Favourites

Places where translation is not possible are still:

- Conversations; translation isn't permitted on statuses with `direct`
visibility
- Search results; more work has to be done there

To do this the primary adjustments are to `TimelineViewModel` and
`NetworkTimelineViewModel`. Previously `TimelineViewModel` contained
code that more properly belonged in `CachedTimelineViewModel`; that code
has been moved there, and the actions code in
`{Cached,Network}TimelineViewModel` is now more explicit.

This also fixes a long-standing TODO in `TimelineFragment`, which was
updating adapter data directly instead of relying on the repository data
to change and invalidate the repository.

Similar changes were made to the `Notifications{Timeline,ViewModel}`
files.

Fixes #359