fix: Prevent "Record not found" when opening as a different account (#1431)
Previous code raced when opening as a different account. In one case the
code would:

1. Open MainActivity
2. Start to change the active account
3. Finish changing the active account
4. Search for the status to open
5. Load the status

In the bad case it would:

1. Open MainActivity
2. Start to change the active account
3. Search for the status to open
4. Finish changing the active account
5. Load the status

I.e., steps 3 and 4 are swapped. When that happens the search is
performed **before** the active account is changed, so the results are
for the previous account. Then (step 5) the new account tries to show
results from the old account. If they are on different servers you get a
"Record not found" error.

While I'm here, `Redirect` et al for the name of the action are
misleading, so rename to `OpenAs` et al.