refactor: Use SharedPreferencesRepository properties for all prefs (#1443)
Previous code had a mix of properties on `SharedPreferencesRepository`
and calls to `getBoolean` etc to get the preference values.

Provide properties for all the preferences and replace the `get...`
calls with property access. In some cases non-preference code had to
call `set...`, so those properties are mutable with custom setters.

This keeps all the default values for preferences in one place and
ensures different parts of the code can't accidentally have different
defaults for the same preference.

There are some related notable changes.

First, `Context.openLink` needed to read from shared preferences to
determine whether to use a custom tab or not. This is replaced with
`OpenUrlUseCase` which can be injected where needed, is activity scoped
(to use the correct context), and receives `SharedPreferencesRepository`
as a parameter to make the appropriate decision.

Second, this has some knock-on effects on `NoUnderlineURLSpan`. That's
been modified to take a click listener (as a functional interface for
Kotlin ergonomics)

Third, `UpdateNotificationFrequency` is moved to `core.preferences` so
it can be used as an `EnumPreference` to simplify other code.