feat: Lab support for exporting/importing preferences
Provide two new lab preference items that export and import the users
preferences.

Preferences are exported as JSON, and include the user's shared preferences
and a redacted copy of the per-account information; enough to restore the
per-account preferences, but without any keys, tokens, etc.

When exporting the user can choose where to save the exported file, and
final file name. When importing the user can try and import any JSON file,
with appropriate error checking.

Export/import is managed by a new `ExportedPreferencesRepository`.