fix: Allow filtering of mentions and severed relationships on older devices (#1454)
There's a custom NotificationPreferencesFragment for devices where
notification preferences aren't handled by the OS.

It was missing options to disable notifications for mentions and severed
relationships, so add them.

The text on the options wasn't capitalised, and was too long for a
one-line preference in some cases, so fix that too.