fix: Set initial reading position for non-cached timelines to the top
This was earlier behaviour, but by accident. b08a0ba28218 broke this when
database lookups were added in `NetworkTimelineViewModel.getStatuses()`.

The data didn't need to be used -- simply making the database lookup was
enough to lose the race. This could also be dependent on network timing and
the speed the remote server returned results, so never showed up in local
tests.

This changed the timing for when results were returned, resulting in
`state.anchorPosition` being much lower down the list (i.e,. instead of
being e.g., `2` it might be `39`.