fix: Don't error on nullable `Report.actionTakenAt` property (#1474)
Two bugs here.

The first is that `Report.actionTakenAt` wasn't marked as nullable when
it should be.

The second is the adapter that converts from JSON to `Instant` wasn't
written to allow null values.

Together these caused notification loading to fail and display an error.

Fix both of those issues.

Fixes #1469