feat: Improve the display of admin notifications
Allow two lines of top text for `admin.report` notifications, so there's
space for the reporter and reportee's details.

Show the comment attached to the report (if present).

Use plural strings so the count of attached posts is grammatically correct.

Ensure the content doesn't run off the edge of the screen in the constraint
layouts.

Don't duplicate the account's name When displaying `admin.signup`
notifications. Previously there was a header line that showed the name,
and a second line that showed the name and the username. Now show the name
(bolded) on the notification top line, and just the username next to the
avatar. Since the same view is used for follow notifications the change
also happens there.