refactor: Simplify *Timeline{ViewModel,Repository}
Remove `*ViewModel.getStatuses`; instead, trigger the fetch by collecting from
`*ViewModel.statuses`, which is flatmapped from a flow that contains the
PachliAccountId to use.

This made it possible to remove the `AccountManager.activeAccountFlow`
references.

The `*TimelineRepository` implementations are modified to only need a
PachliAccountId instead of a full account, which simplifies the code further.