fix: Show translated posts / media descriptions when viewing threads
Previous code didn't fetch the translation state or data from the database
when viewing a thread, so statuses that were translated on a timeline were
not translated in a thread. This also applied to any translated media
descriptions.

Fix this by fetching the local translation info and merging with the network
data, and by ensuring translated media descriptions are passed in onViewMedia.

Fixes #1483