feat: Make media descriptions more important when composing
In the previous code editing the description for a media attachment when
composing the message required:

- Tapping the attachment
- Choosing the "Edit caption" menu item in the popup
- Editing the caption in the dialog that appeared
- Closing the dialog

There was a yellow icon at the top of the screen to indicate the description
was missing.

Change that, adjusting the UI to make the soft-requirement for a media
description more obvious.

UI changes are:

- Attachments are now shown in a vertical list instead of horizontal.
- The extra horizontal space in each list item is used to show a text entry
field for the description.
- Missing descriptions are warnings, not errors. Use Pachli orange as
the warning colour. This removes the need for a warning icon at the top of
the screen.
- The server's media description limit is used, instead of hardcoding it to
1,500 characters.
- The number of characters entered / available is shown.
- Any errors uploading the attachment are shown in the list item immediately,
instead of requiring the user tap the attachment.

Additional functionality changes:

- Previously the description was sent to the server when the user closed the
caption dialog, and an error was shown if it was over-length. Now the
description is sent when the status is sent, and using the server limit
means it's not over-length.