refactor: Move core.data.model.InstanceInfo to core.model
The previous location meant the code for mapping from the network type to
the entity type to the model type was unergonomic, with duplicate code and
constants, and was difficult to follow.

Changes include:

- Move `InstanceInfo` to `core.model`.
- Remove `InstanceInfo.from()`, replace with `.asModel()` functions on the
entity classes, and `.asEntity()` functions on the network classes.
- Make some types on `InstanceInfoEntity` non-null, since the underlying values
should never be null. This requires a database version bump.
- Use existing constants as defaults in the network classes instead of magic
numbers.