refactor: Implement suspendable dialogs (#1499)
In a coroutine it's useful to be able to `await` a dialog and receive
the result, instead of using callbacks.

Ahead of migrating existing dialogs to use `DialogFragment`, create the
`SuspendDialogResult` interface. This provides an `await` function and
`result` property that allows dialogs implementing this interface to
return a result (of arbitrary type) when the dialog is dismissed.

Use this to re-implement the account chooser dialog. Update the calling
code to use the new dialog.