refactor: Extract intent routing to a separate module and activity (#1496)
Previous code sent all intents through `MainActivity`. This had a large
and complicated block of code at the top to figure out where to route
the intent.

Some intents changed the active account, were forwarded elsewhere, and
ended `MainActivity`. Others might redirect elsewhere, to return to 
`MainActivity` when finished. Others were entirely handled by
`MainActivity`.

This had a few problems:

- Complicated to understand and debug if problems with the code
- Made `MainActivity` more complicated when investigating other problems
- Parts of the UI could flash while MainActivity was changing accounts

Fix this be extracting the code that routes intents and determines the
correct account to the new `IntentRouterActivity`.

`IntentRouterActivity` is now the app entry point, responsible for
inspecting the intent, determining which account to use, possibly making
that the active account, and then launching the correct activity to
handle the intent.

This also means that every activity apart from `IntentRouterActivity`
should be launched with an `intent.pachliAccountId` that references a
specific account; using `-1` to indicate the active account should no
longer happen.

Other benefits:

- The splashscreen is retained until loading account data has completed.
- Improved error handling if errors occur when logging in / out. The
idea of a "fallback account" to return to if logging in fails is gone,
the user is always prompted to choose an account.

Fixes #1294, #1316