feat: Provide an "Account Management" UI
Available from the "accounts" switch on the left navigation menu, the UI
shows all accounts, and allows the user to logout individual accounts
(even if the server is no longer available) and switch accounts.

Accessibility actions have also been added for these.

The UI also shows the per-account information about notification fetches,
duplicating information currently in the Preferences UI. That information
might permanently move to the account management UI.

As well as the UI changes, this also:

- Distinguishes between the `Account` and `CredentialAccount` network
types
- Records whether the user's account has been marked as bot, so this can
be shown in the account UI.
- Moves some shared strings to `core.ui`.
- Moves some shared extension functions to `core.ui`.
- Moves `ic_mention_24dp.xml` to `core.ui`.

Fixes #1255