refactor: Cancel Glide requests in fragments earlier (#1519)
Glide is lifecycle aware; when using `Glide.with(someContext)` Glide
will pause any in-progress loads when the lifecycle associated with the
context is stopped, and cancel the load if the lifecycle is destroyed.

Previous code used `Glide.with(someView)...` in a lot of places, causing
Glide to use the view's context and lifecycle.

A view's context is the same as the context of the activity, even if the
view is in a fragment inside an activity. So a load started like this
would only be affected when the containing activity is stopped /
destroyed, not when the fragment is stopped / destroyed.

Fix this by modifying adapters and viewholders to take an explicit
`glide` parameter. This is created with the appropriate activity or
fragment in the caller.

Add a lint detector for this to ensure it doesn't recur.