refactor: Use a Snackbar instead of Bottomsheet when resolving URLs (#1535)
Previous code required anything that used `viewUrl` to contain a view
contain a view that could be used as a bottom sheet with a specific ID.
Messages during URL resolution were shown there.

This seems needlessly complicated; having a specific view is restrictive
and it meant URLs couldn't be resolved in `onCreate`, only in
`onViewCreated` or later.

Replace this with equivalent functionality that displays the message in
a `Snackbar`. This can be called from `onCreate` onwards and doesn't
require a specific view to exist.