fix: Don't leak MainActivity in DrawerImageLoader
The drawer library's `DrawerImageLoader` can leak the object it's initialised
with. This is problem if it's initialised with the activity's `RequestManager`
(Glide) as it previously was, as that holds a reference to the activity, so
the activity is leaked for a period of time.

Fix this by not passing in `RequestManager`, and using `Glide.with(View)` in
this instance.