fix: Don't crash if ComposeActivity is destroyed while taking a picture (#1537)
Previous code saved the URI to save camera images to, but didn't restore
it early enough. If `ComposeActivity` was destroyed and recreated
`onActivityResult` could be called before `photoUploadUri` was restored
(so still null), resulting in an NPE.

Move the state restoration code earlier in `onCreate` so it's guaranteed
to have run.